Finished building the Interfacing layer and tested in with a sub system

+++ Pi/src/comms/model/Response.java
@@ -1,8 +1,7 @@
 package comms.model;
 
-
 public interface Response extends Command {
 
-	public Long getResponseTime();
+	public int getResponseTime();
 
 }
